+++ Source/CesiumRuntime/Private/CesiumGltfComponent.cpp
@@ -1715,6 +1715,12 @@ static void loadPrimitive(
   LODResources.bHasReversedIndices = false;
   LODResources.bHasReversedDepthOnlyIndices = false;
 
+#if ENGINE_VERSION_5_5_OR_HIGHER
+  // UE 5.5 requires that we do this in order to avoid a crash when ray tracing
+  // is enabled.
+  RenderData->InitializeRayTracingRepresentationFromRenderingLODs();
+#endif
+
   primitiveResult.meshIndex = options.pMeshOptions->meshIndex;
   primitiveResult.primitiveIndex = options.primitiveIndex;
   primitiveResult.RenderData = std::move(RenderData);