+++ Source/CesiumRuntime/Public/UnrealAssetAccessor.h
@@ -26,7 +26,7 @@ public:
       const std::string& verb,
       const std::string& url,
       const std::vector<CesiumAsync::IAssetAccessor::THeader>& headers,
-      const gsl::span<const std::byte>& contentPayload) override;
+      const std::span<const std::byte>& contentPayload) override;
 
   virtual void tick() noexcept override;